requirements-dev.txt
feedparser
pygit2
pyyaml
semver
